fully working with proxy on filtering
The 'proper designator sorting' is actually quite difficult to
implement. Before I was dropping data to QStandardItemModel, but because
of correct sorting the filter proxy had to be implemented, and basically
all the operations over the cells now have to be casted through this
proxy. Hence own derivation of QStandardItemModel is not needed any more
on behalf of filtering proxy, and now it is the proxy, which accepts the
drops.