sch-parser at least parses multiple schematics with sheets
tested on umcdrive. But this project revealed serious problems in my
thinking: THE SCHEMATICS ARE HIERARCHIC, and first - the amount of
totally ordered components not only depends on how many of them is in
a single sheet, but as well on how many times the schematic is used in
the hierarchy, and how deep in the hierarchy the schematic is
buried!. Second - these components have always the same designator,
but the fact that they are used many times does not mean, that they
always use the same component, as the design can be repeated, but with
different components for different channels. So this rather calls for
'flattening' the design by taking into consideration the components
hierarchies. Hence some sort of prefix has to be involved in all
this.

The hierarchic schematic walk-through is already added here, but it
only adds the files in order of their appearance. In terms of
designator they are currently indistinguishable!