support for component cache saving and loading
this is quite large change: basically, any time user drop a known
component definition into the designator(s), the dropped component is
paired with libref/value/footprint. Then is is json encoded and
written into componentcache file. If there's already the same
component, no action is taken. The component cache is loaded at the
startup and (will) serves as a 'default' component provider when
context menu is invoked.

The rationale behind this functionality is, that when one e.g. orders
100nF components for one project, he usually orders a complete reel,
or he tends to stick with the same components all the time (unless
some contraindication). Hence it is advantageous to have sort of
'preferred components', those which have been already used in the past
and can be reused.

This greatly simplifies searching for components as those which have
been chosen in the past will be available from the component
cache. Two problems are associated:

- there is _NO FEEDBACK_ on whether by chance supplier pushed the
  particular component from his web pages. it is up to user to check
  whether the information is still correct

- user has an interest to keep the component cache on some global
  place and _under git (or some other) revision_. The filename
  including directory is setup in the global config file. If it does
  not exist, the project directory is taken as base, and the cache is
  stored in componentCache.json of the project
  directory. Currently (with this commit) no way how to setup the
  directory is provided