mimics attribute modifications such, that it generates minimum change
the point of this commit is, that when we save the schematics in kicad,
open in bomizator, save in bomizator, we want to see in git diff
_minimum change_ and only this, which was given by data manipulation of
bomizator (e.g. adding attributes).

it turned out to be quite difficult, as kicad freely enters into the
text spaces at its will (especially in numeric arguments), and it is not
really clear when these are appropriate. Hence this thing generates some
difference indeed, but limited to very small ones, as e.g. one space
somewhere missing in component.

In the future this will have to be clarified how the SCH file is really
coded as e.g. AR attribue is undocumented either.