slowly backporting to standard model
the point is, that proxy should be used only to display the data, and
standarditemmodel should be used to keep the data and take care about
their change. so i started to design this change. it is challenging
and i mostly implemented everything in proxy, what makes issue when
manipulating manually the data