SCH parser now uses normalised designator as key and becomes dictionary
that because it was pain to parse through all the components in a list
while knowing that we can identify it by simple designator. Normalised
designator is a comma separated join of all designators for particular
component, sorted according to 'standard' rules,
i.e. C1,C2,C10,C11,C20... and not basic comparison rule. This for the
moment only loads the schematic and properly shows