radiospares url search, but malfunctioning
it correctly fills-in the radiospares search dialog, but it does
always find nothing. and in order to find something one needs to press
the search button in the web page itself. the text gets however
pre-filled OK. radiospares pages are not user-friendly done at all and
I guess it will be difficult to predict their behaviour