handling of disabled components now part of schparser
schparser needs to process all the data from the schematic, but in
addition to get the project data it needs to load other default data
from a file. Ini file with .bmz is used to get those data. They are
loaded and saved along with the load/save of the project and becomes
'standard' way how to store project specific data. As well added
multiplier, which is for the moment not used